backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.9.2
gTTS==2.5.3
